+++ tests/nativeapp/test_post_deploy_for_app.py
@@ -21,22 +21,20 @@ from pydantic import ValidationError
 from snowflake.cli.api.project.definition_manager import DefinitionManager
 from snowflake.cli.api.project.errors import SchemaValidationError
 from snowflake.cli.api.project.schemas.native_app.application import (
-    ApplicationPostDeployHook,
+    PostDeployHook,
 )
 from snowflake.cli.plugins.nativeapp.exceptions import MissingScriptError
 from snowflake.cli.plugins.nativeapp.run_processor import NativeAppRunProcessor
 
 from tests.nativeapp.patch_utils import mock_connection
 from tests.nativeapp.utils import (
+    CLI_GLOBAL_TEMPLATE_CONTEXT,
     NATIVEAPP_MANAGER_EXECUTE,
     NATIVEAPP_MANAGER_EXECUTE_QUERIES,
     RUN_PROCESSOR_APP_POST_DEPLOY_HOOKS,
 )
 from tests.testing_utils.fixtures import MockConnectionCtx
 
-CLI_GLOBAL_TEMPLATE_CONTEXT = (
-    "snowflake.cli.api.cli_global_context._CliGlobalContextAccess.template_context"
-)
 MOCK_CONNECTION_DB = "tests.testing_utils.fixtures.MockConnectionCtx.database"
 MOCK_CONNECTION_WH = "tests.testing_utils.fixtures.MockConnectionCtx.warehouse"
 
@@ -68,7 +66,7 @@ def test_sql_scripts(
     with project_directory("napp_post_deploy") as project_dir:
         processor = _get_run_processor(str(project_dir))
 
-        processor._execute_post_deploy_hooks()  # noqa SLF001
+        processor.execute_app_post_deploy_hooks()
 
         assert mock_execute_query.mock_calls == [
             mock.call("use database myapp_test_user"),
@@ -115,7 +113,7 @@ def test_sql_scripts_with_no_warehouse_no_database(
     with project_directory("napp_post_deploy") as project_dir:
         processor = _get_run_processor(str(project_dir))
 
-        processor._execute_post_deploy_hooks()  # noqa SLF001
+        processor.execute_app_post_deploy_hooks()
 
         # Verify no "use warehouse"
         # Verify "use database" applies to current application
@@ -148,7 +146,9 @@ def test_missing_sql_script(
         processor = _get_run_processor(str(project_dir))
 
         with pytest.raises(MissingScriptError) as err:
-            processor._execute_post_deploy_hooks()  # noqa SLF001
+            processor.execute_app_post_deploy_hooks()
+
+        assert err.value.message == 'Script "scripts/missing.sql" does not exist'
 
 
 @mock.patch(RUN_PROCESSOR_APP_POST_DEPLOY_HOOKS, new_callable=mock.PropertyMock)
@@ -167,7 +167,7 @@ def test_invalid_hook_type(
         processor = _get_run_processor(str(project_dir))
 
         with pytest.raises(ValueError) as err:
-            processor._execute_post_deploy_hooks()  # noqa SLF001
+            processor.execute_app_post_deploy_hooks()
         assert "Unsupported application post-deploy hook type" in str(err)
 
 
@@ -181,8 +181,8 @@ def test_invalid_hook_type(
 def test_post_deploy_hook_schema(args, expected_error):
     if expected_error:
         with pytest.raises(ValidationError) as err:
-            ApplicationPostDeployHook(**args)
+            PostDeployHook(**args)
 
         assert expected_error in str(SchemaValidationError(err.value))
     else:
-        ApplicationPostDeployHook(**args)
+        PostDeployHook(**args)