+++ tests/nativeapp/test_v2_to_v1.py
@@ -125,9 +125,11 @@ from tests.testing_utils.mock_config import mock_config_key
                     "package": {
                         "name": "pkg_name",
                         "distribution": "external",
-                        "scripts": [
-                            "scripts/script1.sql",
-                            "scripts/script2.sql",
+                        "role": "pkg_role",
+                        "warehouse": "pkg_wh",
+                        "post_deploy": [
+                            {"sql_script": "scripts/script1.sql"},
+                            {"sql_script": "scripts/script2.sql"},
                         ],
                     },
                     "application": {