+++ RELEASE-NOTES.md
@@ -18,9 +18,13 @@
 ## Backward incompatibility
 
 ## Deprecations
+  * Added deprecation warning for `native_app.package.scripts` in project definition file.
 
 ## New additions
 * Added support for project definition file defaults in templates
+* Added support for `native_app.package.post_deploy` scripts in project definition file.
+  * These scripts will execute whenever a Native App Package is created or updated.
+  * Currently only supports SQL scripts: `post_deploy: [{sql_script: script.sql}]`
 
 ## Fixes and improvements